fix mobile tooltip bug
add initial delay to label tooltip.  This fixes an issue on mobile
where pressing the hovering search icon on a utility would trigger the
tooltip for the label appearing on the next page.  This is a workaround
because I couldn't figure out why the tooltip was triggering.  Feel free
to investigate further and implement a proper fix :)